requirements.txt
# HTTP & Async
httpx>=0.27.0
aiohttp>=3.9.0

# Social Media APIs
tweepy>=4.14.0
praw>=7.7.0

# Web Scraping
beautifulsoup4>=4.12.0
lxml>=5.1.0
playwright>=1.40.0

# LLM APIs
anthropic>=0.39.0
openai>=1.50.0

# Config & Environment
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Templating
jinja2>=3.1.0

# CLI
click>=8.1.0

# Logging
rich>=13.7.0

# RSS Parsing
feedparser>=6.0.0

# Date/Time
python-dateutil>=2.8.0
